Organize tests into separate folders and namespaces
