+++ BGPViewerCore.UnitTests/BGPHeServiceTests/BGPHeServiceTest.cs
@@ -1,8 +1,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using BGPViewerCore.Service;
+using Xunit;
 
-namespace Xunit
+namespace BGPViewerCore.UnitTests.BGPHeServiceTests
 {
     public class BGPHeServiceTest : IClassFixture<BGPHeMockWebDriver>
     {
